chore: move implementation of print foreign call into `nargo`
