+++ tooling/nargo/src/foreign_calls/mod.rs
@@ -2,11 +2,11 @@ use std::path::PathBuf;
 
 use acvm::{acir::brillig::ForeignCallResult, pwg::ForeignCallWaitInfo, AcirField};
 use mocker::MockForeignCallExecutor;
-use noirc_printable_type::ForeignCallError;
 use print::{PrintForeignCallExecutor, PrintOutput};
 use rand::Rng;
 use rpc::RPCForeignCallExecutor;
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 
 pub(crate) mod mocker;
 pub(crate) mod print;
@@ -62,6 +62,24 @@ impl ForeignCall {
             _ => None,
         }
     }
+}
+
+#[derive(Debug, Error)]
+pub enum ForeignCallError {
+    #[error("No handler could be found for foreign call `{0}`")]
+    NoHandler(String),
+
+    #[error("Foreign call inputs needed for execution are missing")]
+    MissingForeignCallInputs,
+
+    #[error("Could not parse PrintableType argument. {0}")]
+    ParsingError(#[from] serde_json::Error),
+
+    #[error("Failed calling external resolver. {0}")]
+    ExternalResolverError(#[from] jsonrpc::Error),
+
+    #[error("Assert message resolved after an unsatisified constrain. {0}")]
+    ResolvedAssertMessage(String),
 }
 
 #[derive(Debug, Default)]
